Expired Access_Token and Invalidating Refresh_Token

+++ client/src/pages/category.tsx
@@ -36,7 +36,9 @@ const Category = () => {
 
   const handleDelete = (id: string) => {
     if(!auth.access_token) return;
-    dispatch(deleteCategory(id, auth.access_token))
+    if(window.confirm('Are you sure to delete this category?')){
+      dispatch(deleteCategory(id, auth.access_token))
+    }
   }
