+++ server/controllers/categoryCtrl.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from 'express'
 import Categories from '../models/categoryModel'
+import Blogs from '../models/blogModel'
 import { IReqAuth } from '../config/interface'
 
 const categoryCtrl = {
@@ -60,7 +61,15 @@ const categoryCtrl = {
       return res.status(400).json({msg: "Invalid Authentication."})
 
     try {
+      const blog = await Blogs.findOne({category: req.params.id})
+      if(blog) 
+        return res.status(400).json({
+          msg: "Can not delete! In this category also exist blogs."
+        })
+
       const category = await Categories.findByIdAndDelete(req.params.id)
+      if(!category) 
+        return res.status(400).json({msg: "Category does not exists."})
 
       res.json({ msg: "Delete Success!" })
     } catch (err: any) {
